improved findWords() to work with any String[]

+++ WordItem.java
@@ -1,12 +1,12 @@
 public class WordItem {
   String word;
-  long timeStamp;
+  int timeStamp;
 
-  public WordItem(String word, long timeStamp){
+  public WordItem(String word, int timeStamp){
     this.word = word;
     this.timeStamp = timeStamp;
   }
-  
+
   public String toString(){
     return "Word: " + word + " Time Stamp: " + timeStamp;
   }
